CoreAngular/ClientApp/src/app/counter/counter.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import {NgReduxTestingModule} from '@angular-redux/store/lib/testing';
import { CounterComponent } from './counter.component';
import { CounterActions } from '../redux/actions';

describe('CounterComponent', () => {
  let element: HTMLElement;
  let component: CounterComponent;
  let fixture: ComponentFixture<CounterComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ CounterComponent ],
      imports: [ NgReduxTestingModule ],
      providers: [CounterActions]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CounterComponent);
    component = fixture.componentInstance;
    element = fixture.debugElement.nativeElement;
    fixture.detectChanges();
  });

  it('should create an instance', () => {
    expect(component).toBeTruthy();
  });

  it('should display a title', async(() => {
    const titleText = fixture.nativeElement.querySelector('h1').textContent;
    expect(titleText).toEqual('Counter');
  }));

  it('should start with count 0, then increments by 1 when clicked', async(() => {
    const countElement = fixture.nativeElement.querySelector('strong');
    expect(countElement.textContent).toEqual('0');

    const incrementButton = fixture.nativeElement.querySelector('button');
    incrementButton.click();
    fixture.detectChanges();
    expect(countElement.textContent).toEqual('1');
  }));

  afterEach(() => {
    element.remove();
  });
});
